chore(suite): small SuiteLayout refactor
