+++ packages/suite/src/components/suite/Preloader/SuiteLayout/CoinjoinBars/CoinjoinStatusBar.tsx
